refactor(core,schemas): relocate the profile type defs
relocate the profile type defs
